backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
pydantic
celery
redis
django-storages
boto3
weasyprint
drf-spectacular
python-dotenv
anthropic